+++ __tests__/unit/lib/std.spec.ts
@@ -111,6 +111,7 @@ import {
   ScrollbarFilter,
   LegendHighlight,
   Poptip,
+  TreemapDrillDown,
 } from '../../../src/interaction';
 import {
   SpaceLayer,
@@ -329,6 +330,7 @@ describe('stdlib', () => {
       'interaction.sliderFilter': SliderFilter,
       'interaction.scrollbarFilter': ScrollbarFilter,
       'interaction.poptip': Poptip,
+      'interaction.treemapDrillDown': TreemapDrillDown,
       'composition.spaceLayer': SpaceLayer,
       'composition.spaceFlex': SpaceFlex,
       'composition.facetRect': FacetRect,